fix: tighten warp marker playback design
